backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.27.0
sqlalchemy==2.0.27
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.2.0
openpyxl==3.1.2
